backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.4
aiosqlite>=0.19.0
alembic>=1.13.0

# AI / LLM (Google Gemini - Free Tier)
google-genai>=1.0.0

# Document Parsing
pypdf2>=3.0.0
python-pptx>=0.6.21

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.24.0
